+++ tests/runner/test_multimodal_manager.py
@@ -152,11 +152,9 @@ class TestMultiModalManager:
 
         # Verify the pixel values tensor passed to the mock
         passed_pixel_values = kwargs_arg['pixel_values']
-        assert isinstance(passed_pixel_values, jax.Array)
+        assert isinstance(passed_pixel_values, torch.Tensor)
         assert passed_pixel_values.shape == (1, 3, 224, 224)
-        np.testing.assert_array_equal(
-            np.asarray(passed_pixel_values[0].astype(jnp.float32)),
-            dummy_pixel_values.to(torch.float32).numpy())
+        assert torch.equal(passed_pixel_values[0], dummy_pixel_values)
 
     def test_execute_mm_encoder_multiple_images(self):
         import torch
@@ -258,14 +256,10 @@ class TestMultiModalManager:
         assert "pixel_values" in kwargs_arg
 
         passed_pixel_values = kwargs_arg['pixel_values']
-        assert isinstance(passed_pixel_values, jax.Array)
+        assert isinstance(passed_pixel_values, torch.Tensor)
         assert passed_pixel_values.shape == (2, 3, 224, 224)
-        np.testing.assert_array_equal(
-            np.asarray(passed_pixel_values[0].astype(jnp.float32)),
-            px_1.to(torch.float32).numpy())
-        np.testing.assert_array_equal(
-            np.asarray(passed_pixel_values[1].astype(jnp.float32)),
-            px_2.to(torch.float32).numpy())
+        assert torch.equal(passed_pixel_values[0], px_1)
+        assert torch.equal(passed_pixel_values[1], px_2)
 
     def test_gather_mm_embeddings_chunked_prefill(self):
         """Tests _gather_mm_embeddings with chunked prefill scenarios."""
