+++ tpu_inference/models/vllm/vllm_model_wrapper.py
@@ -29,7 +29,7 @@ import vllm.envs as vllm_envs
 from flax.typing import PRNGKey
 from jax.sharding import Mesh, NamedSharding, PartitionSpec
 from torchax.interop import jax_view, torch_view
-from torchax.ops.mappings import TORCH_DTYPE_TO_JAX
+from torchax.ops.mappings import TORCH_DTYPE_TO_JAX, t2j
 from vllm.config import VllmConfig, set_current_vllm_config
 from vllm.forward_context import set_forward_context
 from vllm.ir import enable_torch_wrap
@@ -108,6 +108,49 @@ class _VllmRunner(torch.nn.Module):
 
     def compute_logits(self, hidden_state: torch.Tensor) -> torch.Tensor:
         return self.vllm_model.compute_logits(hidden_state)
+
+
+class GridTHW(tuple):
+    """Tensor-like wrapper for image/video grid_thw arguments.
+
+    - tuple subclass so isinstance(x, tuple) is True — passes vLLM's
+    tensor_schema type check (e.g. https://github.com/vllm-project/vllm/blob/9744b699bafed423909ed10da96b80eb0542424b/vllm/model_executor/models/qwen3_vl.py#L2026). 
+    - Implements a minimal tensor-like API (ndim, shape, tolist, prod) expected by vLLM's
+    _process_image_input (https://github.com/vllm-project/vllm/blob/9744b699bafed423909ed10da96b80eb0542424b/vllm/model_executor/models/qwen3_vl.py#L2072)
+
+    We cannot use torch.Tensor[tuple] because jax.jit would complain.
+    """
+
+    def __new__(cls, values):
+
+        def _nested_to_tuple(v):
+            if isinstance(v, (list, tuple)):
+                return tuple(_nested_to_tuple(x) for x in v)
+            return int(v)
+
+        flat: tuple = _nested_to_tuple(values)
+        return super().__new__(cls, flat)
+
+    # ---- tensor-like API expected by _process_image_input ----
+
+    @property
+    def ndim(self):
+        return 2
+
+    @property
+    def shape(self):
+        return (len(self), 3)
+
+    def tolist(self):
+        return [list(row) for row in self]
+
+    def prod(self, dim=-1):
+        if dim in (-1, 1):
+            return np.array([row[0] * row[1] * row[2] for row in self])
+        raise NotImplementedError(f"GridTHW.prod({dim}) not supported")
+
+    def __repr__(self):
+        return f"GridTHW({tuple(self)})"
 
 
 class VllmModelWrapper:
@@ -421,29 +464,42 @@ class VllmModelWrapper:
             params_and_buffers: Any,
             **kwargs,
         ) -> Any:
+            call_kwargs = {
+                k: jax.tree.map(torch_view, v)
+                for k, v in kwargs.items()
+            }
 
+            output_from_torch = torch.func.functional_call(
+                self.model,
+                torch_view(params_and_buffers),
+                kwargs={
+                    "call_method": "embed_multimodal",
+                    "call_args": (),
+                    "call_kwargs": call_kwargs,
+                },
+                tie_weights=False,
+            )
+
+            return jax_view(output_from_torch)
+
+        def embed_multimodal_func_torch(params_and_buffers: Any,
+                                        **kwargs) -> Any:
+            # embed_multimodal_func_jax requires kwargs to be jax.Array such that jit can work
+            # Here we move_to_jax, then call (maybe jit'ed) embed_multimodal_func_jax.
             with torchax.default_env(), enable_torch_wrap(False):
 
                 call_kwargs = {
-                    k: jax.tree.map(torch_view, v)
+                    k:
+                    GridTHW(v.tolist()) if k in ("image_grid_thw",
+                                                 "video_grid_thw", "grid_thw")
+                    else jax.tree.map(lambda t: t2j(t, use_dlpack=False), v)
                     for k, v in kwargs.items()
                 }
+                return maybe_jit_embed_multimodal_func(
+                    embed_multimodal_func_jax,
+                    self.vllm_config)(params_and_buffers, **call_kwargs)
 
-                output_from_torch = torch.func.functional_call(
-                    self.model,
-                    torch_view(params_and_buffers),
-                    kwargs={
-                        "call_method": "embed_multimodal",
-                        "call_args": (),
-                        "call_kwargs": call_kwargs,
-                    },
-                    tie_weights=False,
-                )
-
-                return jax_view(output_from_torch)
-
-        return maybe_jit_embed_multimodal_func(embed_multimodal_func_jax,
-                                               self.vllm_config)
+        return embed_multimodal_func_torch
 
     def wrap_embed_input_ids_func(self):
         if not self.vllm_config.model_config.is_multimodal_model:
